Split api into routers, added users, hashed psw, db clean up
