+++ app/database.py
@@ -2,7 +2,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-SQLALCHEMY_DB_URL = 'postgresql://postgres:********@localhost/fastapi'
+SQLALCHEMY_DB_URL = 'postgresql://postgres:Basstard@localhost/fastapi'
 
 engine = create_engine(SQLALCHEMY_DB_URL)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
