+++ app/utils.py
@@ -1,0 +1,6 @@
+from passlib.context import CryptContext
+
+pwd_context = CryptContext(schemes=['bcrypt'], deprecated='auto')
+
+def hash(password: str): 
+  return pwd_context.hash(password)
